backend/requirements.txt

fastapi      
uvicorn      
sqlalchemy   
pydantic     
alembic
psycopg2-binary
python-dotenv
passlib==1.7.4
bcrypt==4.1.2 